+++ CHANGELOG.md
@@ -9,6 +9,8 @@ changes may not be included if they are not expected to break existing code.
 
 ## 0.7.3 (2026-03-19)
 - Add `is_hidden`, `is_unmappable` to `OneSchemaTemplateColumn` type.
+- Use JSON round-trip to clone event payloads, to support Vue's reactive proxy
+  objects.
 
 ## 0.7.2 (2025-07-01)